+++ pkg/engine/grid/column/column.go
@@ -2,5 +2,5 @@ package column
 
 type Column interface {
 	Type() Type
-	Bytes(size int) ([]byte, error)
+	Bytes(size int64) ([]byte, error)
 }
